MAINT: Move files back to the publication place
